+++ src/services/Customers.php
@@ -14,7 +14,7 @@ use Stripe\Customer;
 use Stripe\Invoice;
 use yii\base\Component;
 use enupal\stripe\records\Customer as CustomerRecord;
-
+use Craft;
 
 class Customers extends Component
 {
@@ -104,4 +104,24 @@ class Customers extends Component
 
 		return $customerRecord;
 	}
+
+
+    /**
+     * @param $id
+     * @param $params
+     * @return mixed
+     * @throws \Exception
+     */
+    public function updateStripeCustomer($id, $params)
+    {
+        StripePlugin::$app->settings->initializeStripe();
+        try{
+            Customer::update($id, $params);
+        }catch (\Exception $e){
+            Craft::error('Unable to update Stripe Customer: '.$e->getMessage(), __METHOD__);
+            return false;
+        }
+
+        return true;
+    }
 }